+++ src/main/java/svenhjol/charmony/runestones/common/features/stone_circles/LootTables.java
@@ -1,0 +1,10 @@
+package svenhjol.charmony.runestones.common.features.stone_circles;
+
+import net.minecraft.resources.ResourceLocation;
+import svenhjol.charmony.runestones.RunestonesMod;
+
+public final class LootTables {
+    public static final ResourceLocation STONE_CIRCLE_OVERWORLD_CHEST = RunestonesMod.id("chests/stone_circle_overworld");
+    public static final ResourceLocation STONE_CIRCLE_OVERWORLD_ARCHAEOLOGY = RunestonesMod.id("archaeology/stone_circle_overworld");
+    public static final ResourceLocation STONE_CIRCLE_STONE_RUNESTONE_ARCHAEOLOGY = RunestonesMod.id("archaeology/stone_circle_stone_runestone");
+}